Add drops to Dark Menagerie Mobs
Infested Villager: 10% chance to drop an Emerald when killed by a
player.
Fire Ant: Drops between 0 and 3 String.
Fire Ant: 10% chance to drop a Blaze Powder when killed by a player.
Energy Ghost: 10% chance to drop a diamond when killed by a player.
Energy Ghost: .001% chance to drop a Diamond Sword when killed by a
player.